Fix behaviour of "const_array" method.
GCC complains about the variable 'r' being unused. Lo and behold, it
really isn't used, and would appear to be accidentally discarded. The
method name indicates that it's suposed to return an array of domain d
that all map to value v; but the expr we return returns the const value,
not the array. Fix to return the array.