Mark a ton of Z3 methods as inline.
This is necessary because otherwise they'll generate strongly linked
symbols, and a mess of multiple definition messages wherever we compile
it.

Ideally they'd provide this code as a cpp file and a header that
declares it, but MS in their wisdom have decided otherwise. I'd strip
the two parts into two different files, but then it'd be 100% impossible
to maintain.