Start migrating towards C++api Z3.
This works with regression tests because I've disabled the C++'s api
creation of a reference-counting context. We can now happily muck about
with Z3 C++ objects, migrating towards the new regime, and then turn
reference counting on once everything has been migrated.

(And we'll have to cross our fingers that nothing is going to go wrong,
which it inevitably will do)