Record a z3_convt-wide native int sort.
Right now we generate one at runtime depending on whether or not
int_encoding is true or false. Doing this makes us faster, as well as
making more things use new Z3++ objs.

Also, it gets around a certain problem: you can't create an expr, ast,
sort, whatever to be uninitialized. There isn't a constructor for it.
Thus, we can't have situations like the ones deleted where one copies
onto the top of a sort pointer/object in an outer scope.

We'd have to replace that with a heap allocation (bad bad bad) or hack
uninitialized asts into the c++api, which is probably bad too. The
change in this commit is better than either of those.