Switch a couple of members to z3::sort's.
The purpose was to start using z3::context::array_sort, but this
requires storing the result in a sort obj, and all kinds of turmoil
occurs as a result. All the changes here are largely to do with making
addr_space_tuple_sort and addr_space_arr_sort z3::sort's.