Remove int sort responsibility into z3++.
That is, we make the z3 context wrapping thing aware of what mode we're
in, and have that pre-generate a relevant int sort. This is necessary
due to the number of z3++ api methods that use ctx->int_sort() to get
their integer sort, which wouldn't necessarily match the z3_convt
opinion on the sorts.