Switch tack on how we handle "int_sort"s.
Previously I was of the opinion that we should be hacking ESBMC code
into all the routines that return an "int_val" or int_sort, or whatever.
However then there's ambiguity about when we're getting an int, a bv, or
whatever.

So better to just leave the int_sort and bv_sort stuff lying around in
it's original and then map our own set of esbmc calls on top. These can
take care of whether or not we're in int_encoding mode or not.

Start this off in this patch by declaring an esbmc_int_sort() returning
routine.