Add esbmc-specific int creation methods.
In the case of we're in integer encoding mode, it'll create an int
sorted integer no matter what width we're given. If an int is created
without specifying a width, it'll get the default, a machine width int.
If there's a specified width, it returns a bv_val of it.