Add default, uninitialized Z3 obj constructors.
Problem: As mentioned in the history here, we're unable to declare an
uninitialized z3::expr, and pass a reference to it down to a function
that then initializes it via copy construction. This is because the
objects always want at least a context / something to work with, so no
default empty constructors.

To get around this, add said constructors. That eliminates the property
that all asts/objects contain /something/, but it's not something I see
as valuable within esbmc.

I've inspected the copy constructor in z3::ast, it should work just
fine, no need to add additional null checking.