Move tuple_update and tuple_select to z3_convt
There's no C++api equivalent it appears; and there won't be one unless
we make it, because there are no c++api objects that correspond to a
tuple or any other complicated datatype.

This isn't a bad thing; just I don't want to preserve z3_capi, so move
its contents to z3_convt.