Change how pointer sorts are handled.
Rather than creating a pointer sort whenever we need it, instead set it
up in construction. Also create a func_decl that allows us to easly
generate tuples of its type.

This involves a large amount of flailing in a number of functions to
change them to be all z3::expr'y; ideally it'd have happened over more
than one commit, never mind.