Do some elaborate C++y api twiddling.
We need to fetch the int sort here via the C++ api. However the C++ api
gives no method to use Z3_mk_tuple_sort.

So, put the esbmc int sort into a z3::sort, and then into the array of
Z3_sorts. This way it at least has one reference on the call stack for
the duration of this function, so it's "correct".