Move z3_literal to Z3++ api.
I'm worried that the object returned by z3_literal won't have the
lifetime of a function call, in that:

Z3_mk_something(z3_literal(l));

Will create a z3::expr, referencing the literals AST, then emit the
Z3_ast for Z3_mk_something, then be destroyed, dereferencing (and
freeing) the AST. Which is then passed into Z3_mk_something.

To get around this, I've extended the scope of all the outputs from
z3_literal to not cause this potential problem; If I modify things in
the future that trigger this, then that'll just have to be debugged.