Remove a needless switch of int encoding mode.
These are set by parseoptions, not something else.

Reason for attacking this is because there are a number of places right
now which are switching from the (default) of Z3-BV to Z3-IR, which
isn't valid in my opinion. And it's triggering int2bv, and that's
triggering Z3 4.0 reference counting failures.