Fix counter example printing.
Fetch data from model in a reasonable manner; and actually hand the
right ast to the data-fetching-function.

Of interest is hte model_completion option to Z3_model_eval, which
allows for generating arbitary values whenever there's an undefined
value. Aka, where we currently have gaps, no gaps.

This would be nice for fields in structs which are unimportant but good
for completeness, however would be bad on account of arrays being
possibly vastly huge.