Fix a reference counting fail.
I left this set of calls as it was because I didn't want to touch it; as
it turns out, Z3 reposesses the asts it just allocated via some kind of
internal cache in the course of each function call that creates a new
AST. Therefore it's _imperative_ that something gets reference counted
the very moment it's created, not just eventually. Can't even compose
two Z3_mk_* function calls.

This fixes 13_ludcmp_new