Move all Z3 objects to not be pointers.
That is, the context, various pointer things, are now objects rather
than pointers, making them a little easier to use. Particularly the
pointer function declarations.

This exposes the continuing "get is const" problem: the get method on
prop_convt/z3_convt is marked const, but in reality it needs to do
various non-const things to extract data. Rather than munge that in this
commit, I've made this directory permissive, fix in the next commit.