Fix parsing error in a regression test.
Basically I think what this is looking for is two free input variables,
but that's not what it'll end up with (free pointer != free variable)
and that's not a valid signature for main. So, emulate what the author
probably wanted.

The test still wants this to fail. I don't know why. CBMC can't find a
bug in this either.