Change slight implementation of thread switching.
So, right now running with --smt-during-symex and --smt-thread-guard is
failing, because some code like this:

  IF !(num == 0) THEN GOTO 3
  pthread_cond_wait_check(&empty, &m)

Where num is global, and nonzero, results in a context switch point
occuring with a false state guard. This happens because the next
instruction to be executed has a false guard. Which gets picked up by
--smt-thread-guard as meaning an unviable interleaving.

In reality, it's only an unviable interleaving if we switched from a
point having a false state guard. Which breaks horribly in the above
case, because we're techincally switching from the start of the next
instruction.