Completely bail out of rewriting thread guards.
It's just too painful, complicated, and messes with the symbolic
execution's attempts to merge states upon exiting a conditional.

Instead, share a global state guard across all threads, and guard all
assertions and assumptions with it. This way we get to keep the use of
the state guard, and the global state guard, and they're both
independant of each other. Joy.