Document byte ops; type of expr is now data width.
So there's currently the option of making the type of a
byte_extract/byte_update operation indicate the amount of data to
extract; or to keep byte_* operating on a single byte, and stitch data
together by concatenating them.

The latter is the approach taken by CBMC; it also introduces a byte
concatonate irep. The significant downside of this this approach is that
it adds additional complication to assignments; currently if you assign
to a byte_extract, it becomes a byte_update. But if you're doing it
byte-by-byte, all kinds of jiggery pokery must occur at the symex step.

With the type-is-extract-width approach, we don't get that, but with the
downside that we have additional code in the solver for handling cases
where extracts/updates are split across field boundries, and other fun
with non-deterministic offsets.

IMO the latter is the best place to put the complexity; it's more
concise for longer across more portions of code we have. It's also more
controllable at the SMT layer in terms of efficiency.