Make value set store a _byte_ offset.
That is, whenever we encounter an offset into some kind of address,
calculate a byte size that corresponds to how much the addition is for
in bytes, and then add that to the stored offset. If it's a nondet
offset, that accumulates.

This will break absolutely everything in one or two moments - there's
still the code that _uses_ the offset to hack at.