Reject many dereferences with nonzero offsets.
Essentially, if we start getting nonzero offsets in places, it becomes
massively difficult for us to work out whether or not it can become a
member/index from another fixed location. So, reject them and turn them
into byte updates/selects.

The one exception is selecting from an array where the offset is a
multiple of the array size. We can actually cope with that, as an index.
We could also cope with nondet indexes that are guarenteed to be a
multiple of the element size, but the infrastructure isn't in place
right now to cope with that.