Remove memory_model_conversion.
All this was doing was introducing a new irep for fixedbv conversions,
and typecasting ints of the same size and zero offset into the other
type. That's fine, but all of the exceptional cases could be handled by
memory_model_bytes.

So, pick out that typecast situation immediately, and return. If it's
not that situation, memory_model_bytes it.