Adjust behaviour of type2t::get_width.
From now it returns the size of the field in the byte model. Perhaps it
should also return byte sizes in the future rather than bits, but that's
not important.

What's important is that bools are going to be modelled as single bytes
from now on, to allow them to be properly addressable in our byte
representation model. It might be something we change in the future, but
it needs to be consistent for now.