Resolve some pointer subtypes that are symbols.
When we're performing pointer arithmetic, we have to multiply the
integer part by the size of the pointer it's operating on... that means
working out the size of the subtype of a pointer. Unfortunately, all the
type rewriting goo doesn't rewrite subtypes (because it can be
infinitely recursive). So we have to look up the type at encode time, as
it may be a type symbol.

This involves adding the infrastructure to store a reference to the
namespacet object down to z3_convt.