Start implementing extracting from dyn arrays.
This portion of code works out how many elements to convert, and sets up
selecting its data into the part array.

I haven't gone further because what we should really do, is what the
other portion of the function does: generate an array of data for that
element, and then merge it into the larger part array. Seeing how that's
already implemented, I'm going to move it into a seperate function to
avoid duplication.