Re-engineer the label map
So, this existed in the past so that I could label subformula of some
equation and then be able to see them broken up when the SMT gets
printed to output during debugging. That was great.

Now it occurs to me that this is a great way to link up a piece of SMT
that we've built up, to a symbol, that can then be operated on in terms
of expr2tc's. Possibly. I've since thought up a better solution to my
problem, but this is a good feature anyway.