Convert pointers at _end_ of byte extracts.
So that there's the opportunity for, if we extract 100% of data from a
pointer that's aligned, and the extract type is that of a pointer, we
can just return the pointer. Which would be massively more efficient
than casting it to bits and then back again.

A little more juggling is required for that to happen though.