Clobber renaming of pointers to different type.
This would appear to be the source of much of our grief: whenever we
assign an address of to a pointer, it may in the future get renamed back
to the original object. Which means that the type of the intermediate
pointer gets discarded, and pointer arithmetic goes to pot.

This patch prevents that by not allowing constant propagation if we're
assigning one pointer to another, with different types. The left hand
side then can't be renamed, which would mess with its type in the
future.

I haven't completely clobbered these assignments because there may be
scope for very simple address-of and dereferences being constant
propagated in a way that allows us to cut out all kinds of code. Those
cases should be kept.

However this also isn't perfect; the amount of type renaming of each
pointer type may differ, in which case it'll fail safe and not allow the
propagation.