Fix a dereference insanity.
See deleted code; it would appear that part of the calculations of what
the pointer offset was from an expression, involved subtracting the base
address of the object itself?

Happily that comment was incorrect, it would appear that it just took
the pointer offset of the base object, which is always zero. However,
this needlessly muddies the waters of working out what's going on, so
delete it.