Fix a regression test.
This is blatently wrong: the start addres is set via --function to be
picosat_print, with no definition of it's arguments. Thus, all hell
breaks loose when the file pointer is dereferenced, and we rightly have
an assertion failure.

Because this test isn't supposed to fail, I assume that the code is
correct, but the way we're feeding an argument into picosat_print is
wrong. So, define a free FILE struct on the stack and let the function
use that instead.

--function must die.