Fix yet more pointer arithmetic.
As correctly guessed when I wrote this file, the identified point was
somewhere where pointer arith was insufficiently considered. We do in
fact need to multiply the hoisted constant by the size of the pointer
subtype. So do that. (fixes 03_nec27).

Note that this gives up if it can't get the subtype size out of the
subtype; this might be valid if it hasn't been type renamed enough. If
that happens, it'll be resolved at SMT time, where the renaming will
occur correctly. Kills some constant propagation in the meantime.