Allow for malloc allocations of zero bytes.
Previously what happened, was that we asserted the end of an allocated
region was above the start, to prevent wraparound. However, this can't
be true if the allocated size is zero, so if you malloc'd 0,  you got an
unsat formula. Fix this.

Also allow for malloc returning NULL if the argument is zero.