Workaround dereferencing of multi-dim arrays.
I previously took the approach of thinking the entirety of CBMC was
wrong, and hacked at the C language part of this; turns out it's only
the dereferencing (I think).

See 6.3.2.1 of C89. The dereferencing of something that points at an
array is fundementally some pointer arith, and the dereference itself
performs no action. The spec works through an example. An attempt to
emulate this is implemented by this patch; dereferencing something to an
array type just results in coercing the result of the arith, not in
actually dereferencing anything.