Allow address-ofing a byte extract.
This can occur when we do some pointer arithmetic calculating hte
address of a multi-dimensional array, and the dereferencing code decides
that there's no guarentee that our current offset is element aligned,
thus it produces a byte extract.

In that case if we take the address of it, then we can just undo the
byte extract and use the address of the symbol it's extracting from.
(And it should only ever be extracting from a base symbol or similar).