Disambiguate the use of "rename".
I.E., if we're giving it a number to work with, then we're renaming it
to a specific renaming level. For that, change the method name to
rename_to.

This is because I'm about to twiddle renaming to prevent things like
taking the address of constant_array2t's and suchlike.