Add "no_const_prop" argument to renaming.
The idea being that when we rename something that we're taking the
address of, we prevent any symbols within it being constant propagated
to being a constant array or a constant string. Because that situation
is sucky.

In those situations, we want the address of the variable that they've
been assigned to; for example:

int array[10][10] = { blah };

There we want the address of array, not the address of the data in blah.