Commit an attempt at making byte extract faster.
Basically, the idea is to only encode the minimum number of array
indexes and so forth to make the whole thing workable. Right now it kind
of works, the primary problem is that I need to encode some kind of
guard ensuring that the indexes actually select to the right offset.

Or something like that, I just have to switch branches in a hurry right
now.