Munge throw/catch situation to work with new irep.
This involves throwing in a large number of const_cast's. No comment.

It also involves adding an irep to represent an "ellipsis" expression:
this is so that we can represent something like:

try {
  blah;
} catch (...) {
  throw;
}

Right now I'm not certain where the list of caught expressions actually
comes from, or what irep sits in a CATCH instruction, because
code_cpp_catch2t isn't migrated.