Fix dereferencing C++ class pointers.
We can point a superclass pointer at a base class just fine, and as we
access everything by member _names_ it shouldn't result in any kind of
overlap as a result. (Object comparisons may be wonky, but I can't help
that, it depends how the class is laid out.)

So, search the list of base classes for the target struct type, and
approve the dereference if it's in the list.