Don't zero-initialize external symbols.
Instead, leave them free. Technically if they're accessed at all, that's
an illegal C program because there's no definition of the symbol, and
thus it doesn't have any storage allocated. However, TACAS'13 is really
really sinning in this respect, so go for the broadest intepretation and
make undefined symbols nondet.