Nurk compile-time leaked-memory check system.
It can't cope with conditional allocations, at all, because it might end
up with the variable it's operating on being free (due to the
allocation path not being taken). It's also very unclear to me whether
or not it can cope with multiple allocations from the same position.

So instead, I'm going to move it to symex time munging.