Make value set tracking a little more nuanced.
So: it would appear that the current value set tracking code only has a
few variables that it considers when analysing a particular function.
These are the set of global variables, and the set of local variables,
including function arguments.

Unfortunately C isn't just limited to that - you can also access, via
pointers, the local variables in calling functions. So, it's going to be
necessary for the pointer tracking to be able to track assignments to
caller variables, which doesn't happen right now.

It appears the reason why this data is unavailable is because whenever a
function call occurs, the only data that's shared across the function
boundry is the global variables, and any arguments or return values.
Which is less than ideal. So instead of this, we need to be passing
information down from callers into the value tracking state of the
callees.

At the same time we don't want the local state of callees cluttering up
the value tracking state of the callers, because otherwise every single
instruction in the program will be tracking the state of all possible
pointers in the program, not just the ones it might access.

So the fix: add an option to the 'merge' method of the static analysis
that tells it whether we want to allow non-global non-local variables to
be leaked between state. We then allow this to occur on all internal
transitions of a function, and on transitions from a callers to the
callee, but not the other way around.