Add a guard field to byte operations.
So I was expecting the solver to not interpret symbols that don't
directly contribute to the satisfying model. Unfortunately, because it
has to satisfy every assertion encoded, it has to assign a value to
every lhs of a pointer dereference. That means I can't detect which
dereferences can be un-deferred.

So instead, encode this guard field into the ireps. Then at solve time
look at the model and un-defer the derefs that are enabled in the picked
execution path.