Start collecting a list of global variables.
So, a large amount of solver complexity appears to be sourced from the
initialization of huge global arrays of goo. As a result, accessing the
first element of such an array has to pass through a huge stream of AST
before finding a match.

Perhaps there's a good way of fixing this in SMT (for example, making
such fields lazily made, or making each element an assertion). But Z3
supports arrays with default values, so use them for now.

Do this by extracting a list of all global variables, ignore constraints
applied to them, and add our own that use the Z3 default facilities.