Disable "typeof" instruction for the moment.
To correctly migrate this, I have to interpret the meaning of

  typeid_code.set("type", function.op0().op1().op0().find("#cpp_type"));

Which seems to be dubious at best. I'm also of the opinion that, if this
is implementing C++'s typeof() operator, this should be implemented in
the same way as a vtable, thus making a type lookup just a pointer
dereference.