Commit botched refactor of container-typedef code
The idea here was that we proide relevant information to a macro, that
then creates the expr_methods and something2tc typedefs required for an
irep.

Unfortunately I made some bad decisions, including that ireps always
have data, and that all their superclasses are theirname##_data. Which
is wrong.

So even more typedefs are sprayed around to ensure I could complete the
initial change to using macros; this is then going to need fixing
further.