Add a hack for construction of symbol2tc's.
We can:

  expr2tc(new symbol2t(type, irep_idt_name));

Or:

  expr2tc(new symbol2t(type, basename, lev, l1, l2, trd, node));

To make a symbol2t, however something2tc can only extract the latter
constructor from the set of template arguments that's given to it. So to
make the former constructor available via something2tc, encode a
specific constructor that's only enabled if the container is containing
a symbol2t, and which takes the relevant arguments.