Clean up an ambiguity in not2tc consruction.
If I say:

  not2tc blah(thing);

With thing being an expr2tc, do I mean that I want to upcast+reference
thing in the blah object, or that I want to construct a new irep with
the 'not' operator, applied to thing?

We can't leave this as both, so pick the latter, as I figure the use
cases for upcasting to not2tc (of all things) are fairly limited. And
it's not difficult to use to_not_expr() or whatever