Fix some kind of typing horror.
It seems that even though these two kinds of objects have different
types, lower down in the assignment to tmp_op1 they're passed through an
intermediate type of mul2tc?

The evidence for this is that the upcasting assertion fired, based on
the tmp_op1 line. And it doesn't when the type of mul/ptr_offset is
demoted. Weird, but not a problem now.