Reorder template parameters.
So I'm currently attempting to work out how to rearrange the
list_operands goo to be better. Right now it's seriously inefficient in
that it performs a series of allocs and frees in frequent code paths.
Code should just be able to iterate over exprs without causing a
penalty.

The plan for now is to reorder templates so that all the fields that
contain expr2tc's in some shape or form come first. Then we can have a
get_expr(unsigned i) that fetches the relevant field, null if it's not
an expr2tc, upon which the loop stops

And I shall try this shortly.