Fix a change in list_operands2 behaviour.
Basically, we can now get empty ireps when listing operands. This is
entirely valid, because an empty irep is a valid operand to another irep
(and presumably it means something).

We could skip them, but I want future code to be able to use the
indicies from get_sub_expr and make them consistent for "a while".