Fix a massive brain fail.
We should only mark something as changed if we actually successfully
simplified a sub operand of it. It looks like, when I originally made
the get_sub_expr change, I moved the assign to changed into the most
likely looking is_nil_expr block.

Wheras actually we need two (is the sub expr null; is the simplified
expr null)