Elaborate on SMT conversion design.
This fleshes out a little of the idea of converting all expressions into
applications of SMT functions, or terminals. I've limited the number of
arguments to 3 as I don't know of any functions that use more; there's
no use in using a dynamic number of arguments because all it results in
is another heap allocation for the same amount of wasted accounting
memoryh.