Revert back to no information in smt_{ast,sort}.
As explained in the comments, on third thoughts I don't see any reason
for solver-independant data be tracked here. The there shouldn't be any
case where the smt code needs to look at this stuff in a solver
independant way. Solvers themselves can dump dat in an extension of this
class if they like.

The exception is debugging, but there should be enough data tracked in
the opaque class to facilitate debugging things.