Start storing an expr2tc in z3 smt asts.
The idea here is that there's some mixture of SMT ASTs and z3 internal
converted gunge. The problem being that we have the maximum amount of
information available right now during conversion, so that we can fall
back to the old method and keep all tests passing all the way through
conversion.