Skip thinking about pointer arith for now.
We reach the inevitable problem that it's possible to express pointer
arithmatic in the SMT backend, and that should be converted down to
something else before it hits the solver.

However I want to have a few more functions supported before really
tackingling that, so go for the default solution for the moment.