Start thinking about generic SMT conversion.
i.e., there are going to be a bunch of expr kinds that can be converted
to functions without any real logic, and only need some very light
assertions. So, we can probably automated it in a table.

This means that we're going to have to have a variety of hooks in there
to ensure things like we don't use BV operations in integer mode, and
that we can catch pointer arithmetic before it gets converted.