Store sorts in smt_asts.
We're going to end up needing this to guess at the validity of exprs
being converted, i.e. whether their operands match up with what we
expect.

Additionally, it should (in the end) end up with sign data being
available in the smt conversion portion of code. It's starting to appear
like that's going to be the case.