Fix a failure mode involving pointer validity.
Basically, when infinite sized arrays are used, there's no SMT way of
representing that all elements need a default value. In the past this
was handled with Z3's default array value extension.

Clearly this isn't valid SMT, so as we move to an SMT only backend
the fix was to detect all accesses to array elements, and assert that
they're the default value in the initial array. Which works, until you
constant propagate away that original array. Therefore, disallow const
propagation of this.