First shot at defining an smt-only tuple situation
Basically, the idea is to make the essence of a tuple its name. From
that, we have symbols (identified by a suffix) that contain the actual
data of the tuple. So a tuple becomes a collection of symbols. There's
no link between the symbols at the SMT layer aside from having a common
prefix.

Recursive structs are handled by adding more prefix/suffix to the name.