Seriously remunge the 'main object' situation.
Now that we're generating it for more than one place (uuugh, cross
compiling) we instead give the 'common' file the name of the mainobj,
and construct MAINOBJ from that.

This necessitates the 'all:' rule to fall below the inclusion of common.
Yet gmake, being rubbish, interprets the first rule in the file as the
main rule. So we have to link these two by saing that 'all' depends on
'realall'. ugh.