Start munging tuple-array updates.
Take the usual approach: we generate a temporary fresh constant of the
right sort, and bind things into it. Normal sorts right now, we'll see
what happens in the future. Seeing how we're working on tuple arrays,
each element is actually an array.

We'll be ignoring structs for a while.