ast-ify tuple_create_rec.
So, after looking at one of the constant structs that was being
constructed, there was only one argument, a 'member' irep selecting some
kind of pointer value out of another struct. This is perfectly valid C,
you construct a data item on the stack say using expressions that pull
data out of others structs.

However the approach of fetching sub-exprs out of exprs doesn't work in
this circumstance, we actually need to interprete + evaluate the
expressions that are making up these struct members. So, convert
everything to an smt ast first, and then extract data from that point
onwards (which guarentees operational exprs are evaluated).