Move deadlock-check indirection to preprocessor.
I think I avoided doing this in the past on account of either not
knowing that I should just rename pthread_lib.c functions til it worked,
or because I didn't /want/ those names to change in counterexamples, for
example.

However, having deadlock indirection pass through the entire AST proves
to be massively expensive in C++. So don't.