Fix a massive fixedbv-comparison fail.
At the time when I produced some code that was implementing relations, I
figured that integers would be the only things that made their way past.
So to detect signed comparisons, I checked for signedbv types.

Of course, fixedbvs / other non-integer types are signed too. So fix
this to model that.