Fix typecast simplification error.
It would appear that we're supposed to derive the type of a fixedbv from
the target type of the case we're simplifying. Which seems to make
sense. Actually do that rather than call to_fixedbv, which seems to do
the wrong thing.

Given that the current way is broken, in the future to_fixedbv should be
munged to take a target fixedbv spec and use .to_integer, or whatever.
Either way, future work.