Massively hack the irep2 hash calculations.
i.e., keep a copy of the hash value in the actual expr we're hashing.
And only ever change it when someone performs a mutable operation on the
expr, which we can detect via the expr2tc container.

This works because we don't support using exprs outside of containers,
and the fact that we can null out the crc when a mutable operation
occurs, then recalculate it when a const operation occurs.