Scatter various const casts around.
This is so that Clang can compile all of these things. Fundementally,
we're entirely broken by the fact that everyone ignores the
const-correctness of all these types, and someday we're going to end up
with a load of irep attached to the 'nil' irep, and will have no idea
where it came from, because people randomly shoehorn data in here and
there. Bleahg.

(or we could use types...)