Seriously remunge some throwing stuff.
Rather than reaching into other instructions and fiddling with their
jump targets, or storing the thrown value in the global symbol table,
instead actually symbolically assign it, and allow the
goto-state-merging gunk decide when there's a target for a throw
instruction to be run. That really doesn't have to be handled in the
throw/catch logic.

That being said, it's unclear what the has_throw_target field actually
means.