Set guard to false if we fail to handle a throw.
Either this code ends up ending here as an assertion failure, or has a
jump encoded that takes it to the function handling unhandled
exceptions. Either way, execution from this point does not continue.

Also encode a check in symex handling that steps over throws if the
guard is false.