Remunge thrown object handling.
Firstly, the target location where I was deciding whether to merge in
a thrown object was wrong; we should in fact have been pointing two
instructions further, at the assignment to the variable that contains
the thrown object.

Secondly, the check in symex_step that the current stack_catch wasn't
empty is wrong: the assignment that gets merged in is /after/ the end of
the CATCH bracketed code. So if we only have one level of try{}catch{}
then nothing happens.