Disable inlining once and for all.
Ish. It's still enabled if you specify --inlining on the command line,
but that's not going to happen very often. As far as I can tell, this
was basically inlining absolutely everything there was due to some
inverted conditions.

This was breaking regression/esbmc-cpp/esbmc-algorithm/algorithm100
after the irep2 move. I don't know what was keeping it working
pre-irep2, but either way the inlining was causing a variable in the
vector class iterator constructor (pvec_pos) to be assigned twice,
leading to a bunch of double assignments and death. Seems that the
activation record wasn't updated, or whatever.

Inlining: now considered harmful.