Specialise pointer comparisons SMT conversion.
Previously we casted to integer and compared those. That works just fine
and accurately, but fantastically slowly. I suspect that the addition of
an array with followed by some arithmetic lowers the solver to bit
blasting. The test esbmc-cpp/esbmc-algorithm/algorithm101 takes about a
bajillion hours to run as a result of 55737ae, which introduced this
feature.

So instead, decide whether the pointer objects are the same; if they
are, then just compare the offsets. If they're not, then just compare
the start addresses.

This means that we just pass some values around without manipulating
them, which I suspect is exactly what the SMT solver is very good at
dealing with. No arithmetic is introduced at all, which is great.