Massacre inlining renaming situation.
Right now, when we inline, all the local variables to the inlined
functions have their l1 activation record number set to whatever the
function that inlined them's number is. This is clearly broken, leads to
double assignments to the same variable name, making all formulas unsat
(see: regression/esbmc-cpp/esbmc-algorithm/algorithm100).

Fix this by a) tracking what functions are inlined (070bd749) and b)
when we enter a function, bump all the l1 names of inlined functions
seperately. This fixes the aforementioned test. Huurrr.

Unfortunately it may also "break" (TM) a bunch of other tests in that
now they don't get silly answers, but there's not a lot I can do about
that.

This all becomes significantly ugly, particluarly goto_symext::locality.
We should probably look at, instead, storing a set of local variable
names in function records (instead of the instructions, although this
leads to complexity with decl instructions).

Then store a set of pointers to inlined functions in a function, and
avoid having to enumerate + search variable names in
goto_symext::locality.