Replace list of inlined funcs with set.
This is to prevent multiple inlinings resulting in multiple calls to
goto_symext::locality and so forth. If a function is inlined more than
once into a function right now, l2 renaming handles making variables
distinct. It's unclear whether accessing uninitialized variables
evaluates to the previous invocations last variables. (But that's
probably the case).