Refactor the local variable scenario.
As mentioned a little earlier, now extract a list of local variables
from the program and store it somewhere. This is then preserved, so that
we can distinguish inlined variables from non-inlined. (It turns out
that using the variable name to do that is impossible).

This is going to be unpleasent to maintain, as we're factually
duplicating some data, which is generally bad. At some point in the
future though I'm going to blitz the instruction/program representation
and hopefully clarify all the things.