Fix a C++ naming convention situation.
Right now we have memory dereference checks that what we're pointing at
(and dereferencing) is of the right type. Normally this checks that the
types are the same (expected type and actually dereferenced type).
However in circumstances where we have a C++ class being dereferenced as
whatever it's superclass is, that's valid too.

One of the naming conventions in the C++ frontend stopped working, so
fix it up to be the correct convention again.