Fix migration of certain symbols (and printing).
So, beards in my eyes. I was relying on an '@' symbol in a symbol name
meaning that it'd been renamed to at least l1. And now C++ has come
along with @virtual_table and such like, which neatly screws up that
plan. (Not C++'s fault, storing this data as a string is silly in the
first place).

The implications of this are unknown, but I do know it broke the
expr2tct::id_shorthand method test to see whether a name had a base
name, because exprs in instructions were already renamed (due to the
issue above), and thus didn't appear in the symbol table.

It might also be that this has been screwing up even more C++ in the
meantime (although only on the irep2 branch). This would explain a few
things.

Reference test: esbmc-cpp/esbmc-inheritance/dynamic_cast_simple, which
if you run with --show-goto-functions produces some rubbish and renamed
names in the initialization of class vtables. With this patch, that
doesn't happen (and the test doesn't fail).