Substancially alter how l1 renaming works.
So it turns out that l1 renaming was seriously out of kilter in
different places: as we entered a function, we renamed all local
variables (in /that/ scope) to the function invocation number. However
when we entered another block (and subsequent 'decl' instructions) we
just increased the l1 renaming number of variables when they were
redeclared.

That means we could have a function invocation where some variable ('a')
in a for loop gets renamed to have l1 numbers 1 through 5, as it
iterates through the for loop a few times.

Then when the function is re-entered, 'a' gets renamed to 2, because
that's the function invocation number, after which it gets renamed to
have l1 numbers 2 through 6.

That leads to a double assignment, which may very well make an
always-unsat formula.

Fix this by changing the function invocation count to a variable
instance count: that way variable names aren't linked to the funciton
invocation number, and _always_ get an incrementing number when they're
renamed, so there's no chance of regression.

This kind of makes my previous inline-tracking-code aimless, but never
mind, it led to more knowledge.